// client/src/store/types/lockTypes.ts
enum lockTypes {
  INIT_LOCKS = "INIT_LOCKS",

  GET_ALL_LOCKS_START = "GET_ALL_LOCKS_START",
  GET_ALL_LOCKS_SUCCESS = "GET_ALL_LOCKS_SUCCESS",
  GET_ALL_LOCKS_FAIL = "GET_ALL_LOCKS_FAIL",

  OPEN_LOCK_START = "OPEN_LOCK_START",
  OPEN_LOCK_SUCCESS = "OPEN_LOCK_SUCCESS",
  OPEN_LOCK_FAIL = "OPEN_LOCK_FAIL",

  RESET_LOCK_START = "RESET_LOCK_START",
  RESET_LOCK_SUCCESS = "RESET_LOCK_SUCCESS",
  RESET_LOCK_FAIL = "RESET_LOCK_FAIL",

  DELETE_LOCK_START = "DELETE_LOCK_START",
  DELETE_LOCK_SUCCESS = "DELETE_LOCK_SUCCESS",
  DELETE_LOCK_FAIL = "DELETE_LOCK_FAIL",
}

export default lockTypes;